+++ canaille/backends/memory/models.py
@@ -267,7 +267,7 @@ class User(canaille.core.models.User, MemoryModel):
         return bool(self.password)
 
     def check_password(self, password):
-        if password not in self.password:
+        if password != self.password:
             return (False, None)
 
         if self.locked: